requirements.txt
grpcio==1.38.0
grpcio-tools==1.38.0
protobuf==3.17.3
torch==1.9.0
transformers==4.8.2
numpy==1.20.3
redis==3.5.3 